added authorization stripe and cronjobs

+++ beaconinteriorsapi/Services/OrderService.cs
@@ -17,7 +17,11 @@ namespace beaconinteriorsapi.Services
         }
         public async Task<IEnumerable<Order>> GetAllOrdersAsync()
         {
-            return await _dbContext.Orders.Include(o => o.Addresses).Include(o => o.Items).ToListAsync();
+            return await _dbContext.Orders.Include(o => o.Addresses).Include(o => o.Items).Include(o=>o.User).ToListAsync();
+        }
+        public async Task<IEnumerable<Order>> GetUserOrdersAsync(string id)
+        {
+            return await _dbContext.Orders.Include(o => o.Addresses).Include(o => o.Items).Where(o=>o.UserId==id).ToListAsync();
         }
         public async Task<Order> GetSingleOrderAsync(Guid id)
         {
@@ -25,12 +29,37 @@ namespace beaconinteriorsapi.Services
             if (order == null) ThrowNotFound($"no order found with id:{id.ToString()}");
             return order!;
         }
-        public async Task DeleteOrderAsync(Guid id)
+        public async Task<Order> TrackOrderAsync(string id)
+        {
+            var order = await _dbContext.Orders.Include(o => o.Addresses).Include(o => o.Items).FirstOrDefaultAsync(o => o.TrackingID == id);
+            if (order == null) ThrowNotFound($"no order found with id:{id.ToString()}");
+            return order!;
+        }
+        public async Task<bool> CheckIfOrderExpiredAsync(string id)
+        {
+            var order = await _dbContext.Orders.Include(o => o.Addresses).Include(o => o.Items).FirstOrDefaultAsync(o => o.TrackingID == id);
+            if (order == null) ThrowNotFound($"no order found with id:{id.ToString()}");
+            if(order!.HasExpired) return true;
+            else
+            {
+                order.PaymentStatus=PaymentStatusType.Processing;
+                await _dbContext.SaveChangesAsync();
+                return false;
+            }
+        }
+        public async Task DeleteOrderAsync(Guid id,IEnumerable<string> roles,string userId)
         {
             var order=await _dbContext.Orders.Include(o => o.Addresses).Include(o => o.Items).FirstOrDefaultAsync(o => o.Id == id);
             if (order == null) ThrowNotFound($"no order found with id:{id.ToString()}");
-            _dbContext.Orders.Remove(order!);
-            await _dbContext.SaveChangesAsync();
+            if (roles.Any(r => r.Equals(UserRoleType.SuperAdmin.ToString())) || order!.UserId == userId)
+            {
+                _dbContext.Orders.Remove(order!);
+                await _dbContext.SaveChangesAsync();
+            }
+            else
+            {
+                ThrowUnauthorizedError("unauthorized");
+            }
         }
     }
 }
